pyproject.toml
[build-system]
requires = [
    "setuptools>=65",
    "wheel",
    "pybind11>=2.6.1",
]
build-backend = "setuptools.build_meta"

[project]
name = "veccs"
description = "A Python package for efficiently computing conditioning sets for Vecchia approximations"
readme = "README.md"
license = {text = "MIT License"}
authors = [
    {name = "Matthias Katzfuss", email = "katzfuss@gmail.com"},
    {name = "Myeongjong Kang", email = "kmj.stat@gmail.com"},
    {name = "Jian Cao", email = "jcao21@uh.edu"},
    {name = "Paul Wiemann", email = "wiemann.3@osu.edu"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
]
requires-python = ">=3.10, <3.13"
dependencies = [
    "numpy >= 1.24.2",
    "scipy >= 1.10.1",
    "deprecated > 1.2",
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "veccs.__version__.__version__"}

[project.optional-dependencies]

legacy = [
    "faiss-cpu >= 1.7.3",
    "numpy<2.0",
    "scikit-learn >= 1.2.2",
]

dev = [
    "black",
    "flake8",
    "isort",
    "mypy>=0.991",
    "pre-commit",
    "pytest",
    "pytest-cov",
]

docs = [
    "mkdocs>=1.4.2",
    "mkdocstrings[python]>=0.20.0",
    "mkdocs-material>=9.1.2",
    "mkdocs-gen-files>=0.4.0",
    "mkdocs-literate-nav>=0.6.0",
]

[tool.isort]
profile = "black"
py_version = 310

[tool.mypy]
files = ["src", "tests"]
ignore_missing_imports = true
scripts_are_modules = true

# turn warnings into errors
[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore:distutils Version classes are deprecated:DeprecationWarning",
    "ignore:numpy.core._multiarray_umath is deprecated and has been renamed to numpy._core._multiarray_umath",
]

[tool.pyright]
exclude = [ "venv" ]
venvPath = "."
venv = "venv"
